face-recognition-backend-flask/requirements.txt
appdirs==1.4.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
distlib==0.3.2
filelock==3.0.12
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.4.7
Werkzeug==2.1.2
zipp==3.8.0
